requirements.txt
dvc==3.63.0
mlflow==2.22.0
pandas==2.3.0
scikit-learn==1.7.0
joblib==1.5.1
PyYAML==6.0.2
matplotlib==3.10.3
pathspec==0.12.1
feast
psycopg[binary]
psycopg_pool
gdown
tqdm
marimo
diagrams
opencv-python
jupyter
ipykernel
